added implementation for item and product JPA Data Access

+++ src/main/java/com/philips/shoppingcart/repository/ItemJpaRepository.java
@@ -3,6 +3,6 @@ package com.philips.shoppingcart.repository;
 import com.philips.shoppingcart.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemJpaRepository extends JpaRepository<Item,Integer> {
+public interface ItemJpaRepository extends JpaRepository<Item,Long> {
 
 }
