+++ src/main/java/com/philips/shoppingcart/repository/ShoppingCartJpaRepository.java
@@ -3,5 +3,5 @@ package com.philips.shoppingcart.repository;
 import com.philips.shoppingcart.model.ShoppingCart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingCartJpaRepository extends JpaRepository<ShoppingCart,Integer> {
+public interface ShoppingCartJpaRepository extends JpaRepository<ShoppingCart,Long> {
 }
